sensors: split airspeed selector header into sperate file
